+++ tests/test_dynamodb2/exceptions/test_dynamodb_exceptions.py
